10 тз. 2 проверка

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -3,10 +3,14 @@ package ru.yandex.practicum.filmorate.controller;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exceptions.FilmNotFoundException;
+import ru.yandex.practicum.filmorate.exceptions.UserNotFoundException;
+import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.film.FilmService;
 
 import javax.validation.Valid;
+import java.time.LocalDate;
 import java.util.*;
 
 @RestController
@@ -23,13 +27,27 @@ public class FilmController {
     //Создание фильма
     @PostMapping
     public Film create(@Valid @RequestBody Film film) {
-        return filmService.create(film);
+        if (film.getReleaseDate().isBefore(LocalDate.of(1895, 12, 28))) {
+            log.warn("Фильм с id = {} не создан. Дата выпуска фильма не может быть раньше даты создания кино!",
+                    film.getId());
+            throw new ValidationException("Фильм не создан. Дата выпуска фильма не может быть раньше даты" +
+                    " создания кино!");
+        } else {
+            return filmService.create(film);
+        }
     }
 
     //Обновление фильма
     @PutMapping
     public Film update(@Valid @RequestBody Film film) {
-        return filmService.update(film);
+        if (film.getReleaseDate().isBefore(LocalDate.of(1895, 12, 28))) {
+            log.warn("Фильм с id = {} не обновлен. Дата выпуска фильма не может быть раньше даты создания кино!",
+                    film.getId());
+            throw new ValidationException("Фильм не обновлен. Дата выпуска фильма не может быть раньше даты" +
+                    " создания кино!");
+        } else {
+            return filmService.update(film);
+        }
     }
 
     //Получение фильма по id
@@ -47,21 +65,30 @@ public class FilmController {
     //Поставить лайк
     @PutMapping("{filmId}/like/{userId}")
     public void addLike(@PathVariable long filmId, @PathVariable long userId) {
-        filmService.addLike(filmId, userId);
+        if (userId < 0) {
+            throw new UserNotFoundException("id указан меньше 0!");
+        } else if (filmId < 0) {
+            throw new FilmNotFoundException("id фильма не может быть отрицательным!");
+        } else {
+            filmService.addLike(filmId, userId);
+        }
     }
 
     //Удалить лайк
     @DeleteMapping("{filmId}/like/{userId}")
     public void unlike(@PathVariable long filmId, @PathVariable long userId) {
-        filmService.deleteLike(filmId, userId);
+        if (userId < 0) {
+            throw new UserNotFoundException("id указан меньше 0!");
+        } else if (filmId < 0) {
+            throw new FilmNotFoundException("id фильма не может быть отрицательным!");
+        } else {
+            filmService.deleteLike(filmId, userId);
+        }
     }
 
     //Список лучших фильмов
     @GetMapping("/popular")
-    public Collection<Film> getPopularFilms(@RequestParam(required = false) Optional<Integer> count) {
-        if (count.isPresent()) {
-            return filmService.getMostPopularFilms(count.get());
-        }
-        return filmService.getMostPopularFilms(10);
+    public Collection<Film> getPopularFilms(@RequestParam(defaultValue = "10") Integer count) {
+        return filmService.getMostPopularFilms(count);
     }
 }
